SIMD 128: add namespaces according to granularity
Granularity is type of integer that will be operated in parallel in
operations of __m128i numbers.

It includes currently `u16` and `u32` for `uint16_t` and `uint32_t`.
